fix: refactor to allow desktop to work
it would skip to the end since desktop loads 100 at a time whereas
mobile is way less. it now runs through each chunk 10 at a time and also
doesn't remove the watched until it hits the end or the modal is closed.
it will then jump to the top and remove all based on the ratio set.